+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Core: fast reliable network protocol, fast multi-threaded server, utf8 chat, bui
 Changelog
 ======
 
+### 5.15.2.0 (?)
+  * Client MCP support: enable_mcp=1
+  * Mapgen Earth: Add voxel earth ( https://github.com/voxelearth ) voxelizer: voxel_earth=1  ;  voxel_earth_api_key=YOUR_GOOGLE_3D_TILES_API_KEY_HERE
+
 ### 5.14.0.0 (?)
   * Farmesh fixes and settings
      * Works with client_mesh_chunk: Just set client_mesh_chunk=4 for better quality